Fix measureElectricity loop increment
as measureElectricity loops, it will push upstream to AWS/HA iff the measureIndex
is == the config.h LOCAL_MEASUREMENTS (30 by default, or every 30 measurements,
being every 30s). The reset of the measureIndex counter happens outside of the
check that the increment has reached its target, meaning every loop the index is
reset back to 0 and never reaches 30. Move the reset to 0 to the condition that
it reached its target.